deployable_project/requirements.txt
numpy==1.26.4 --only-binary=numpy
Flask==2.0.1
supabase==1.0.3
pandas==1.3.3 --only-binary=pandas
yfinance==0.1.70
pytz==2021.1
python-dotenv==0.19.0
gunicorn==20.1.0
Werkzeug==2.0.1
Flask-Login==0.5.0
Flask-WTF==0.15.1
Flask-SQLAlchemy==2.5.1
Flask-Migrate==3.1.0
Flask-Babel==2.0.0
Flask-Caching==1.10.1
Flask-Mail==0.9.1
Flask-CORS==3.0.10
Flask-JWT-Extended==4.3.1
Flask-RESTful==0.3.9
Flask-SocketIO==5.1.1
Flask-Admin==1.5.8
pytest==6.2.5
pytest-cov==2.12.1
black==21.7b0
flake8==3.9.2
isort==5.9.3
mypy==0.910 